backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
pydantic>=2.5
passlib[bcrypt]
python-jose[cryptography]
python-multipart
psycopg2-binary
reportlab
qrcode[pil]
pillow
python-dotenv
email-validator
google-auth
